correct signals and general control portions count for any planner
