+++ ydb/core/protos/flat_scheme_op.proto
@@ -575,6 +575,7 @@ message TCompactionSelectorConstructorContainer {
 
 message TCompactionPlannerConstructorContainer {
     optional string ClassName = 1;
+    optional uint64 NodePortionsCountLimit = 2;
 
     message TLOptimizer {
 
@@ -588,7 +589,6 @@ message TCompactionPlannerConstructorContainer {
     message TLCOptimizer {
         repeated TCompactionLevelConstructorContainer Levels = 1;
         repeated TCompactionSelectorConstructorContainer Selectors = 2;
-        optional uint64 NodePortionsCountLimit = 3;
     }
 
     message TTilingOptimizer {
