+++ ydb/core/tx/columnshard/engines/storage/optimizer/abstract/counters.cpp
@@ -1,0 +1,5 @@
+#include "counters.h"
+
+namespace NKikimr::NOlap::NStorageOptimizer {
+
+}
