+++ ydb/core/tx/columnshard/engines/storage/optimizer/abstract/ya.make
@@ -2,6 +2,7 @@ LIBRARY()
 
 SRCS(
     optimizer.cpp
+    counters.cpp
 )
 
 PEERDIR(
