+++ ydb/core/tx/columnshard/engines/storage/optimizer/lcbuckets/constructor/constructor.h
@@ -15,7 +15,6 @@ public:
 private:
     std::vector<TLevelConstructorContainer> LevelConstructors;
     std::vector<TSelectorConstructorContainer> SelectorConstructors;
-    ui32 NodePortionsCountLimit = 1000000;
 
     static inline const TFactory::TRegistrator<TOptimizerPlannerConstructor> Registrator =
         TFactory::TRegistrator<TOptimizerPlannerConstructor>(GetClassNameStatic());
