+++ ydb/core/tx/columnshard/engines/storage/optimizer/lcbuckets/planner/optimizer.cpp
@@ -14,14 +14,13 @@ namespace NKikimr::NOlap::NStorageOptimizer::NLCBuckets {
 TOptimizerPlanner::TOptimizerPlanner(const TInternalPathId pathId, const std::shared_ptr<IStoragesManager>& storagesManager,
     const std::shared_ptr<arrow::Schema>& primaryKeysSchema, std::shared_ptr<TCounters> counters, std::shared_ptr<TSimplePortionsGroupInfo> portionsGroupInfo, std::vector<std::shared_ptr<IPortionsLevel>>&& levels,
     std::vector<std::shared_ptr<IPortionsSelector>>&& selectors, const ui32 nodePortionsCountLimit)
-    : TBase(pathId)
+    : TBase(pathId, nodePortionsCountLimit)
     , Counters(counters)
     , PortionsInfo(portionsGroupInfo)
     , Selectors(std::move(selectors))
     , Levels(std::move(levels))
     , StoragesManager(storagesManager)
-    , PrimaryKeysSchema(primaryKeysSchema)
-    , NodePortionsCountLimit(nodePortionsCountLimit) {
+    , PrimaryKeysSchema(primaryKeysSchema) {
     RefreshWeights();
 }
